Made code more readable

+++ modules/user/deleteUser.php
@@ -4,13 +4,23 @@
     include 'class.user.php';
     include '../rbac/class.rbac.php';
 
-    $uid  = $_SESSION['uid'];
-    $rbac = new rbac($uid);
-    $rbac ->checkSession();
-    $rbac ->checkAccess();
-    
-    $id   = $_POST['id'];
-    $user = new User($id);        
-    $user ->deleteUser();        
+    $uid = $_SESSION['uid'];
 
+    $user = new User($uid);
+    $rbac = new Rbac($uid);
+    $role = $rbac->getUserRoleNumber();
+    
+    if(!$uid){
+        header('Location: http://localhost/loginsystem/index.php');
+        exit();          
+    }else if($role > 1){
+        $_SESSION['access'] = 2;
+        header('Location: http://localhost/loginsystem/main.php');
+        exit();               
+    }else{
+        $id   = $_POST['id'];
+        $user = new User($id);        
+        $user ->deleteUser();   
+    }
+     
 ?>
